Extension up and running.
It now is able to scan through the ingredients lists to determine if there is a matching work in the avoid list.

Unfortunately, it can't tell the difference between 'virgin' and 'gin', or 'crumble' and 'crum'.

I need to figure out a way to split each ingredient array into amounts and item so that I can scan the full name of the item.