Risk items now only display once.
Formerly displayed everytime they appeared in a category or as an ingredient, now they only display once.

Fixed a few risk words (such as 'gin' and 'rum') to have spaces on either side so they don't trigger on words such as "virgin" and "crumble". Need to find a cleaner way to avoid this.